requirements.txt
pandas
azure-identity
azure-storage-blob
tenacity
dash
plotly

